refactor: MaterialService methods

+++ src/main/java/com/luminabackend/services/MaterialService.java
@@ -3,17 +3,17 @@ package com.luminabackend.services;
 import com.luminabackend.exceptions.AccessDeniedException;
 import com.luminabackend.exceptions.EntityNotFoundException;
 import com.luminabackend.exceptions.ZipProcessingException;
-import com.luminabackend.models.education.classroom.Classroom;
 import com.luminabackend.models.education.material.Material;
-import com.luminabackend.repositories.material.MaterialRepository;
 import com.luminabackend.models.user.dto.user.UserAccessDTO;
+import com.luminabackend.repositories.material.MaterialRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.io.ByteArrayResource;
-import org.springframework.data.mongodb.gridfs.GridFsResource;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.data.mongodb.gridfs.GridFsResource;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
+
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.util.List;
@@ -37,15 +37,11 @@ public class MaterialService {
     @Autowired
     private ClassroomService classroomService;
 
-    public List<Material> getAllMaterials(UUID classroomId, UserAccessDTO userAccessDTO) {
-        Classroom classroom = classroomService.getClassroomById(classroomId);
-        permissionService.checkAccessToClassroom(classroom, userAccessDTO);
+    public List<Material> getAllMaterials(UUID classroomId) {
         return repository.findMaterialByClassroomId(classroomId);
     }
 
-    public Page<Material> getPaginatedClassroomMaterials(UUID classroomId, UserAccessDTO userAccessDTO, Pageable page) {
-        Classroom classroom = classroomService.getClassroomById(classroomId);
-        permissionService.checkAccessToClassroom(classroom, userAccessDTO);
+    public Page<Material> getPaginatedClassroomMaterials(UUID classroomId, Pageable page) {
         return repository.findMaterialByClassroomId(classroomId, page);
     }
 
@@ -55,16 +51,13 @@ public class MaterialService {
         return materialById.get();
     }
 
-    public ByteArrayResource getAllMaterialsAsZip(UUID classroomId, UserAccessDTO userAccessDTO) throws IOException {
-        Classroom classroom = classroomService.getClassroomById(classroomId);
-        permissionService.checkAccessToClassroom(classroom, userAccessDTO);
-        List<Material> materials = getAllMaterials(classroomId, userAccessDTO);
+    public ByteArrayResource getAllMaterialsAsZip(UUID classroomId) throws IOException {
+        List<Material> materials = getAllMaterials(classroomId);
 
-        try (ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
-             ZipOutputStream zipOutputStream = new ZipOutputStream(outputStream)) {
+        try (ByteArrayOutputStream outputStream = new ByteArrayOutputStream(); ZipOutputStream zipOutputStream = new ZipOutputStream(outputStream)) {
 
             materials.forEach(material -> {
-                if (material.getId() != null){
+                if (material.getId() != null) {
                     GridFsResource gridFsResource = fileStorageService.getFile(material.getFileId());
                     if (gridFsResource != null) {
                         addFileToZip(gridFsResource, zipOutputStream);
@@ -88,44 +81,29 @@ public class MaterialService {
         }
     }
 
-    public Material saveMaterial(
-                                    UUID classroomId,
-                                    UserAccessDTO userAccessDTO,
-                                    String title,
-                                    String description,
-                                    MultipartFile file) throws IOException {
-
-        Classroom classroom = classroomService.getClassroomById(classroomId);
-        permissionService.checkAccessToClassroom(classroom, userAccessDTO);
-
+    public Material saveMaterial(UUID classroomId, UUID professorId, String title, String description, MultipartFile file) throws IOException {
         String fileId = fileStorageService.storeFile(file, classroomId);
-        Material material = new Material(classroomId, userAccessDTO.id(), title, description, fileId);
+        Material material = new Material(classroomId, professorId, title, description, fileId);
         return repository.save(material);
     }
 
-    public void deleteById(UUID materialId, UUID classroomId, UserAccessDTO userAccessDTO) {
-        Classroom classroom = classroomService.getClassroomById(classroomId);
-        permissionService.checkAccessToClassroom(classroom, userAccessDTO);
-
+    public void deleteById(UUID materialId) {
         Material material = getMaterialById(materialId);
 
-        if (material.getFileId() != null)
-            fileStorageService.deleteFile(material.getFileId());
+        if (material.getFileId() != null) fileStorageService.deleteFile(material.getFileId());
         repository.deleteById(materialId);
     }
 
     public void deleteAllByClassroomId(UUID classroomId) {
         List<Material> materialByClassroomId = repository.findMaterialByClassroomId(classroomId);
         materialByClassroomId.forEach(material -> {
-            if (material.getFileId() != null)
-                fileStorageService.deleteFile(material.getFileId());
+            if (material.getFileId() != null) fileStorageService.deleteFile(material.getFileId());
             repository.deleteById(material.getId());
         });
     }
 
-    public void checkAccessToMaterial(UUID classroomId, UUID materialId, UserAccessDTO userAccessDTO){
-        Classroom classroom = classroomService.getClassroomById(classroomId);
-        permissionService.checkAccessToClassroom(classroom, userAccessDTO);
+    public void checkAccessToMaterial(UUID classroomId, UUID materialId, UserAccessDTO userAccessDTO) {
+        permissionService.checkAccessToClassroomById(classroomId, userAccessDTO);
         Material material = getMaterialById(materialId);
         if (!material.getProfessorId().equals(userAccessDTO.id()))
             throw new AccessDeniedException("You don't have permission to access this resource");
